refactor: Enhance type safety in SignIn component and integrate Sidebar in multiple components
